replace month format with short one

+++ application/js/calendar.js
@@ -160,7 +160,7 @@ export default class Calendar {
 	checkMonth(element, nextElement, i){ // adds month to date with month change
 		if (i < 34 && element.dateObj.getDate() > nextElement.dateObj.getDate()) {
 			if (i < 7) {
-				element.month = element.dateObj.toLocaleString('en-GB', {month: 'long'});
+				element.month = element.dateObj.toLocaleString('en-GB', {month: 'short'});
 			} else if (i > 27) {
 				nextElement.month = nextElement.dateObj.toLocaleString('en-GB', {month: 'short'});
 			}
